requirements.txt
Flask==2.3.2
openai==0.27.8
google-generativeai==0.8.3
python-dotenv==1.0.1
ollama